Move triangle_button to widgets
